requirements.txt
conan 
oftest 
pyzmq